Math tasks long --> long long.

+++ source/main/cTaskLib.cc
@@ -2111,15 +2111,15 @@ double cTaskLib::Task_Mult(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
 
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
 
   for (int i = 0; i < input_size; i ++) {
     for (int j = 0; j < input_size; j ++) {
       if (i == j) continue;
-      long cur_diff = labs(long(input_buffer[i] * input_buffer[j]) - test_output);
+      long long cur_diff = labs((long long)(input_buffer[i] * input_buffer[j]) - test_output);
       if (cur_diff < diff) diff = cur_diff;
     }
   }
@@ -2155,15 +2155,15 @@ double cTaskLib::Task_Div(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
     for (int j = 0; j < input_size; j ++) {
       if (input_buffer[j] == 0) continue;
-      long cur_diff = labs(long(input_buffer[i] / input_buffer[j]) - test_output);
+      long long cur_diff = labs((long long)(input_buffer[i] / input_buffer[j]) - test_output);
       if (cur_diff < diff) diff = cur_diff;
     }
   }
@@ -2199,13 +2199,13 @@ double cTaskLib::Task_Log(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(log(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
+    long long cur_diff = labs((long long)(log(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
   
@@ -2241,13 +2241,13 @@ double cTaskLib::Task_Log2(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(log2(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
+    long long cur_diff = labs((long long)(log2(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
   
@@ -2283,13 +2283,13 @@ double cTaskLib::Task_Log10(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(log10(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
+    long long cur_diff = labs((long long)(log10(fabs(double(input_buffer[i] ? input_buffer[i] : 1)))) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
   
@@ -2325,13 +2325,13 @@ double cTaskLib::Task_Sqrt(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(sqrt(fabs(double(input_buffer[i])))) - test_output);
+    long long cur_diff = labs((long long)(sqrt(fabs(double(input_buffer[i])))) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
   
@@ -2367,13 +2367,13 @@ double cTaskLib::Task_Sine(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(sin(double(input_buffer[i]) / dCastPrecision) * dCastPrecision) - test_output);
+    long long cur_diff = labs((long long)(sin(double(input_buffer[i]) / dCastPrecision) * dCastPrecision) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
   
@@ -2408,13 +2408,13 @@ double cTaskLib::Task_Cosine(cTaskContext& ctx) const
   const cArgContainer& args = ctx.GetTaskEntry()->GetArguments();
   
   const tBuffer<int>& input_buffer = ctx.GetInputBuffer();
-  const long test_output = ctx.GetOutputBuffer()[0];
+  const long long test_output = ctx.GetOutputBuffer()[0];
   const int input_size = input_buffer.GetNumStored();
   
-  long diff = (INT_MAX + 1) * 2;
+  long long diff = (INT_MAX + 1) * 2;
   
   for (int i = 0; i < input_size; i ++) {
-    long cur_diff = labs(long(cos(double(input_buffer[i]) / dCastPrecision) * dCastPrecision) - test_output);
+    long long cur_diff = labs((long long)(cos(double(input_buffer[i]) / dCastPrecision) * dCastPrecision) - test_output);
     if (cur_diff < diff) diff = cur_diff;
   }
